Moves messages
