+++ src/loserMessages.ts
@@ -1,0 +1,22 @@
+const LOSER_MESSAGES = [
+  "You rolled a 1, you loser",
+  "More like death ROLLED! GOTTEM!",
+  "That roll was sus! Are you the Imposter?",
+  "You just got death-ROLLED!",
+  "1? That's not gonna cut it, chief",
+  "Oof, that roll's got me DYING!",
+  "We're gonna need a bigger failboat for that giant 1 you just rolled!",
+  "Oops! All 1s.",
+  "That roll? Delete your account.",
+  "1? I think you dropped this 🫴 L",
+  "Is that your IQ or your roll?",
+  "Oh look, you're number 1 at being the worst!",
+  "The '1' stands for '1 more reason you're not winning.'",
+  "You must have a Ph.D. in Onesology, the study of rolling 1s!",
+  "Oh, you rolled a 1? I almost didn't notice among all your other failures.",
+  "Is this your strategy? Because it's a terrible 1.",
+  "You really nailed that roll! Too bad it was into your own coffin. RIP.",
+  "Some people bring joy wherever they go; you bring 1s.",
+]
+
+export default LOSER_MESSAGES
